Move test data creation to TestEnv.kt.
