+++ integration-tests/src/test/kotlin/io/spine/examples/protodata/hello/ApplySizeOptionPluginTest.kt
@@ -171,43 +171,3 @@ class `ApplySizeOptionPlugin should` {
         )
     }
 }
-
-private fun createInvalidContactBuilder(): Contact.Builder {
-    return Contact.newBuilder()
-        .setElementCount(2)
-        .addPhone("Phone")
-        .addEmail("Email")
-        .addAddress(createAddressBuilder(1))
-}
-
-private fun createValidContactBuilder(): Contact.Builder {
-    val elementCount = 3
-
-    val builder = Contact.newBuilder()
-        .setElementCount(elementCount)
-
-    repeat(elementCount) {
-        builder
-            .addPhone("Phone$it")
-            .addEmail("Email$it")
-            .addAddress(
-                createAddressBuilder(it)
-            )
-    }
-    return builder
-}
-
-private fun createAddressBuilder(seed: Int): Address.Builder {
-    val numberOfLines = 2
-
-    val builder = Address.newBuilder()
-        .setNumberOfLines(numberOfLines)
-        .setZipcode("Zipcode$seed")
-        .setCountry("Country$seed")
-
-    repeat(numberOfLines) {
-        builder.addAddressLine("AddressLine$seed$it")
-    }
-
-    return builder
-}
